// api.ts
import { API_KEY } from "@env";
import { QueryFunction, UseQueryResult } from "react-query";
const BASE_URL = "https://api.themoviedb.org/3";

interface MoviesType {
    nowPlaying: QueryFunction;
    upcoming: QueryFunction;
    trending: QueryFunction;
};

export interface Movie {
    adult: boolean;
    backdrop_path: string | null;
    genre_ids: number[];
    id: number;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    release_date: string;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
};
interface BaseResponse {
    page: number;
    total_results: number;
    total_pages: number;
};

export interface MovieResponse extends BaseResponse {
    results: Movie[];
};
export type QueryMovieReturn = UseQueryResult<any, any> & { data?: MovieResponse; };

export const movies: MoviesType = {
    nowPlaying: () =>
        fetch(`${BASE_URL}/movie/now_playing?api_key=${API_KEY}&language=en-US&page=1`).then(res => res.json()), 
    upcoming: () =>
        fetch(`${BASE_URL}/movie/upcoming?api_key=${API_KEY}&language=en-US&page=1`).then(res => res.json()), 
    trending: () =>
        fetch(`${BASE_URL}/trending/movie/week?api_key=${API_KEY}&language=en-US&page=1`).then(res => res.json())
};